fix: stoch acts float()

+++ agents/erode.py
@@ -172,8 +172,7 @@ class Agent(Base):
                 # sample stochastic actions
                 stoch_samples = math.floor(self.cfg.popsize * (1 - self.cfg.mix_coeff))
                 dist = TruncatedNormal(loc=mean, scale=var, a=-2, b=2)  # range [-2,2] to avoid discontinuity at [-1,1]
-                stoch_acts = dist.sample(
-                    sample_shape=[stoch_samples, ])  # output popsize x horizon x action_dims matrix
+                stoch_acts = dist.sample(sample_shape=[stoch_samples, ]).float()
                 stoch_acts = torch.where(stoch_acts < torch.tensor([-1.0], device=self.device),
                                          torch.tensor([-1.0], device=self.device),
                                          stoch_acts)  # clip
